Finished base of code
A lot might be wrong because I cant test it but for now its pretty much done, this was one of the easiest things ive done in python. There was the ability to make this more interesting and complex but for now I wasn't allowed to do any of that.